+++ CHANGELOG.md
@@ -9,6 +9,7 @@
   - Korean (ko): Fix typo in `equal_to` keys #1061
   - Portuguese (pt, pt-BR): add translation for `errors.messages.in` #1071
   - Russian (ru): fix some errors in 'datetime' section
+  - Scottish Gaelic (gd): Add locale
   - Spanish (es): add translation for `errors.messages.in` #1071
   - French (fr, fr-CA, fr-CH, fr-FR): fix typo on 'almost_x_years: one' #1074
 - Add ordinalization for German (de, de-AT, de-CH, de-DE)